cleanup integrations
